move/rename some forge specific stuff
