+++ src/main/java/com/gregtechceu/gtceu/data/DataGenerators.java
@@ -1,4 +1,4 @@
-package com.gregtechceu.gtceu.data.forge;
+package com.gregtechceu.gtceu.data;
 
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.api.registry.registrate.SoundEntryBuilder;
